feat: add roles

+++ backend/microservices/question/src/controllers/question-controller.ts
@@ -1,7 +1,8 @@
 import { Request, Response } from 'express';
+import { StatusCodes } from 'http-status-codes';
 import { Question, Difficulty } from '@/models/question';
 import { QuestionService, CreateQuestionReq } from '@/services/question-service'; 
-
+import { handleServiceError } from '@/controllers/error-handler';
 
 const questionService = new QuestionService();
 
@@ -10,23 +11,30 @@ export async function createQuestion(req: Request, res: Response) {
     const newQuestion: CreateQuestionReq = req.body;
     await questionService.createQuestion(newQuestion);
 
-    res.status(201).send();
+    res.status(StatusCodes.OK).send();
   } catch (error) {
-    res.status(500).send();
+    handleServiceError(error, res)
+    res.send();
   }
 }
 
 export async function getQuestionById(req: Request, res: Response) {
   try {
     const id = req.params.id;
+    if (!id) {
+      res.status(StatusCodes.BAD_REQUEST).json("missing id params")
+      return
+    }
+
     const question = await questionService.getQuestionById(id);
     if (question) {
       res.json(question);
     } else {
-      res.status(404).send();
+      res.status(StatusCodes.NOT_FOUND).send();
     }
   } catch (error) {
-    res.status(500).send();
+    handleServiceError(error, res)
+    res.send();
   }
 }
 
@@ -43,7 +51,6 @@ export async function getFilteredQuestions(req: Request, res: Response) {
         .filter(d => !isNaN(d) && Object.values(Difficulty).includes(d))
         .map(d => d as Difficulty);
     } else {
-      // get all questions
       difficultiesEnumArray = Object.values(Difficulty).map(d => d as Difficulty)
     }
 
@@ -56,8 +63,8 @@ export async function getFilteredQuestions(req: Request, res: Response) {
 
     res.json(sortedQuestions);
   } catch (error) {
-    console.error(error);
-    res.status(500).send('Internal server error');
+    handleServiceError(error, res)
+    res.send();
   }
 }
 
@@ -65,30 +72,42 @@ export async function getFilteredQuestions(req: Request, res: Response) {
 export async function updateQuestion(req: Request, res: Response) {
   try {
     const id = req.params.id;
+    if (!id) {
+      res.status(StatusCodes.BAD_REQUEST).json("missing id params")
+      return
+    }
+
     const update: Partial<Question> = req.body;
     const updatedQuestion = await questionService.updateQuestion(id, update);
 
     if (updatedQuestion) {
       res.json(updatedQuestion);
     } else {
-      res.status(404).send('Question not found');
+      res.status(StatusCodes.NOT_FOUND);
     }
   } catch (error) {
-    res.status(500).send('Internal Server Error');
+    handleServiceError(error, res)
+    res.send();
   }
 }
 
 export async function deleteQuestion(req: Request, res: Response) {
   try {
     const id = req.params.id;
+    if (!id) {
+      res.status(StatusCodes.BAD_REQUEST).json("missing id params")
+      return
+    }
+
     const deleted = await questionService.deleteQuestion(id);
 
     if (deleted) {
-      res.status(204).send();
+      res.status(StatusCodes.NO_CONTENT).send();
     } else {
-      res.status(404).send();
+      res.status(StatusCodes.OK).send();
     }
   } catch (error) {
-    res.status(500).send();
+    handleServiceError(error, res)
+    res.send();
   }
 }
