+++ backend/microservices/user/src/controllers/auth-controller.ts
@@ -1,6 +1,10 @@
-import { Request, Response } from "express";
-import authService from "@/services/auth-service";
-import userService from "@/services/user-service";
+import e, { Request, Response } from 'express';
+import authService from '@/services/auth-service';
+import userService from '@/services/user-service';
+import { StatusCodes } from 'http-status-codes';
+import { handleServiceError } from '@/controllers/error-handler';
+import { JsonWebTokenError } from 'jsonwebtoken';
+
 
 const testEmail = "example@email.com";
 
@@ -35,14 +39,15 @@ export async function googleAuth(req: Request, res: Response): Promise<void> {
       user = await userService.create(testEmail, "");
     }
 
-    const accessToken = await authService.generateAccessToken(user!.id);
-    const refreshToken = await authService.generateRefreshToken(user!.id);
+    const accessToken = await authService.generateAccessToken(user.id, user.roles);
+    const refreshToken = await authService.generateRefreshToken(user.id);
 
     res.cookie(accessTokenKey, accessToken, cookieConfig);
     res.cookie(refreshTokenKey, refreshToken, cookieConfig);
 
-    res.status(200).json(user);
+    res.status(StatusCodes.OK).json(user)
   }
+
   const authUrl = authService.getGoogleAuthURL();
   res.redirect(authUrl);
 }
@@ -62,7 +67,7 @@ export async function googleRedirect(
       user = await userService.create(userInfo.email, userInfo.picture);
     }
 
-    const accessToken = await authService.generateAccessToken(user!.id);
+    const accessToken = await authService.generateAccessToken(user!.id, user.roles);
     const refreshToken = await authService.generateRefreshToken(user!.id);
 
     res.cookie(accessTokenKey, accessToken, cookieConfig);
@@ -70,8 +75,10 @@ export async function googleRedirect(
 
     res.redirect(loginRedirectURL);
   } catch (error) {
-    console.error("Google OAuth callback error:", error);
-    res.status(500).json({ message: "Google Auth failed" });
+    console.error('googleRedirect: ', error);
+    
+    handleServiceError(error, res)
+    res.send();
   }
 }
 
@@ -84,13 +91,19 @@ export async function refresh(req: Request, res: Response): Promise<void> {
 
       const tokenStore = await authService.readRefreshToken(decodedToken.id);
       if (!tokenStore || tokenStore.revoked) {
-        console.log("Invalid refresh token");
-        res.status(500).send();
+        res.status(StatusCodes.UNAUTHORIZED).send()
+      }
+
+      let user = await userService.read(decodedToken.userId);
+
+      if (!user) {
+        res.status(StatusCodes.BAD_REQUEST).send()
+        return
       }
 
       // sync delete, if failed dont continue
       await authService.deleteRefreshToken(decodedToken.id);
-      const accessToken = authService.generateAccessToken(decodedToken.userId);
+      const accessToken = authService.generateAccessToken(user?.email, user?.roles);
       const refreshToken = await authService.generateRefreshToken(
         decodedToken.userId
       );
@@ -98,15 +111,18 @@ export async function refresh(req: Request, res: Response): Promise<void> {
       res.cookie(accessTokenKey, accessToken, cookieConfig);
       res.cookie(refreshTokenKey, refreshToken, cookieConfig);
 
-      res
-        .status(200)
-        .json({ access_token: accessToken, refresh_token: refreshToken });
-    } catch (err) {
-      console.log("Token verification failed", err);
-      res.status(401).send();
+      res.status(StatusCodes.OK).send()
+    } catch(error) {
+      if (error instanceof JsonWebTokenError) {
+        res.status(StatusCodes.UNAUTHORIZED).json({ message: "Invalid Token/Expired token"})
+      } else {
+        handleServiceError(error, res)
+      }
+
+      res.send();
     }
   } else {
-    res.status(401).send();
+    res.status(StatusCodes.UNAUTHORIZED).send()
   }
 }
 
@@ -117,12 +133,12 @@ export async function checkAuth(req: Request, res: Response): Promise<void> {
     try {
       authService.verifyAccessToken(token).userId;
 
-      res.status(200).send();
-    } catch (err) {
-      res.status(401).send();
+      res.status(StatusCodes.OK).send()
+    } catch(err) {
+      res.status(StatusCodes.UNAUTHORIZED).send()
     }
   } else {
-    res.status(401).send();
+      res.status(StatusCodes.UNAUTHORIZED).send()
   }
 }
 
@@ -136,12 +152,18 @@ export async function logout(req: Request, res: Response): Promise<void> {
 
       res.clearCookie(accessTokenKey, cookieConfig);
       res.clearCookie(refreshTokenKey, cookieConfig);
-
-      res.status(200).send();
-    } catch (err) {
-      res.status(401).send();
+          
+      res.status(StatusCodes.OK).send()
+    } catch(err) {
+      res.status(StatusCodes.UNAUTHORIZED).send()
     }
   } else {
-    res.status(400).send();
+      res.status(StatusCodes.OK).send()
+  }
+}
+
+export class OAuthError extends Error {
+  constructor(message: string) {
+      super(message);
   }
 }
